fix: update account deletion webhook and updated reffrer path in monday-connection-check middleware

+++ src/controllers/management-controller.js
@@ -8,23 +8,36 @@ const TAG = "management_controller";
 const connectionService = new ConnectionModelService();
 
 export const deleteAccountAuth = async (req, resp) => {
-  const { accountId } = req.query;
-  await connectionService.deleteConnection(accountId);
-  logger.info("Deleted account auth", TAG, { accountId });
-  resp.status(200).send();
-};
+  const webhookBody = req.body;
 
+  const accountId = webhookBody.data.account_id;
+
+  if (webhookBody.type === "uninstall") {
+    await connectionService.deleteConnection(accountId);
+    logger.info("Deleted account auth", TAG, { accountId });
+    resp.status(200).send();
+  }
+};
 
 export const setAccountFeatureFlag = async (req, resp) => {
   const { flag, enabled } = req.body;
-  const {accountId} = req.query;
-  const mondayConnection = await connectionService.getConnectionByAccountId(accountId);
-  try{
-    const result = await setAccountFlag({_token: mondayConnection.mondayToken}, flag, enabled);
-    resp.status(200).send({result: result});
+  const { accountId } = req.query;
+  const mondayConnection = await connectionService.getConnectionByAccountId(
+    accountId
+  );
+  try {
+    const result = await setAccountFlag(
+      { _token: mondayConnection.mondayToken },
+      flag,
+      enabled
+    );
+    resp.status(200).send({ result: result });
+  } catch (err) {
+    logger.error("Error setting account flag", TAG, {
+      error: err.message,
+      flag,
+      enabled,
+    });
+    return resp.status(500).send({ error: err.message });
   }
-  catch(err){
-    logger.error("Error setting account flag", TAG, {error: err.message, flag, enabled});
-    return resp.status(500).send({error: err.message});
-  }
-}
+};
